Move project participation repo functions to own repo file Add `hasLoaded` property to dataloader update tests
