+++ app/src/hooks/useDataLoader.ts
@@ -28,6 +28,12 @@ export type DataLoader<AFArgs extends any[], AFResponse = unknown, AFError = unk
    */
   isReady: boolean;
   /**
+   * `true` if the `fetchData` function has been called at least one time.
+   *
+   * @type {boolean}
+   */
+  hasLoaded: boolean;
+  /**
    * Executes the `fetchData` function once, only if it has never been called before. Does nothing if called again.
    */
   load: (...args: AFArgs) => void;
@@ -72,7 +78,8 @@ export default function useDataLoader<AFArgs extends any[], AFResponse = unknown
   const [error, setError] = useState<AFError | unknown>();
   const [isLoading, setIsLoading] = useState(false);
   const [isReady, setIsReady] = useState(false);
-  const [isOneTimeLoad, setOneTimeLoad] = useState(false);
+  const [hasLoaded, setHasLoaded] = useState(false);
+  const [oneTimeLoad, setOneTimeLoad] = useState(false);
 
   const isMounted = useIsMounted();
 
@@ -100,11 +107,12 @@ export default function useDataLoader<AFArgs extends any[], AFResponse = unknown
     } finally {
       setIsLoading(false);
       setIsReady(true);
+      !hasLoaded && setHasLoaded(true);
     }
   };
 
   const load = (...args: AFArgs) => {
-    if (isOneTimeLoad) {
+    if (oneTimeLoad) {
       return;
     }
 
@@ -113,9 +121,10 @@ export default function useDataLoader<AFArgs extends any[], AFResponse = unknown
   };
 
   const refresh = (...args: AFArgs) => {
-    setError(undefined);
-    setIsLoading(false);
-    setIsReady(false);
+    error && setError(undefined);
+    isLoading && setIsLoading(false);
+    isReady && setIsReady(false);
+    !hasLoaded && setHasLoaded(true);
     loadData(...args);
   };
 
@@ -124,12 +133,13 @@ export default function useDataLoader<AFArgs extends any[], AFResponse = unknown
   };
 
   const clearData = () => {
-    setError(undefined);
     setData(undefined);
-    setIsReady(false);
-    setIsLoading(false);
-    setOneTimeLoad(false);
+    error && setError(undefined);
+    isLoading && setIsLoading(false);
+    isReady && setIsReady(false);
+    hasLoaded && setHasLoaded(false);
+    oneTimeLoad && setOneTimeLoad(false);
   };
 
-  return { data, error, isLoading, isReady, load, refresh, clearError, clearData };
+  return { data, error, isLoading, isReady, hasLoaded, load, refresh, clearError, clearData };
 }
